Corrections based on feddback
in this new update, i made adjustment to my code based on the feedback i recieved. For example i was able to align my container to the center and i figured out the problem was from my media queries. Also i removed the fixed heights and widths unless they were necessary which mostly weren't.
i also ajdusted the semantic html to simple div instead of article and for the image i placed that in a figure.  For the media queries, i re-ordered them in ascending order.